work in progress signin-signup page, rename root component to login-view
