+++ libs/auth/src/lib/auth.module.ts
@@ -12,9 +12,10 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatIconModule } from '@angular/material/icon';
 
 // Component
-import { AuthRootComponent } from './root/root.component';
+import { LoginViewComponent } from './pages/login-view/login-view.component';
 import { UiFormModule } from '@blockframes/ui';
 import { SignupFormComponent } from './components/signup-form/signup-form.component';
 import { SigninFormComponent } from './components/signin-form/signin-form.component';
@@ -35,9 +36,10 @@ export const AuthRoutes: Routes = [
     MatInputModule,
     MatButtonModule,
     MatSidenavModule,
+    MatIconModule,
     RouterModule.forChild(AuthRoutes),
     UiFormModule
   ],
-  declarations: [AuthRootComponent, SigninFormComponent, SignupFormComponent],
+  declarations: [LoginViewComponent, SigninFormComponent, SignupFormComponent],
 })
 export class AuthModule {}
