+++ libs/auth/src/lib/components/signin-form/signin-form.component.html
@@ -1,5 +1,5 @@
 <section testId="signin" fxLayout="column" fxLayoutAlign="start center">
-  <h1>Sign in</h1>
+  <h2>Sign in</h2>
   <form [formGroup]="signinForm" (ngSubmit)="submited.emit(signinForm)" fxLayout="column">
     <input-email [form]="signinForm"></input-email>
     <input-password [form]="signinForm" [showHint]="false"></input-password>
