+++ libs/auth/src/lib/pages/login-view/login-view.component.html
@@ -1,16 +1,16 @@
 <mat-sidenav-container>
-  <mat-sidenav #signinSidenav [opened]="isSignin" mode="side">
-    <auth-login-form (opened)="isSignin=false" (submited)="signin($event)"></auth-login-form>
-  </mat-sidenav>
   <mat-sidenav-content [fxLayoutAlign]="align">
-    <button class="mat-elevation-z0" mat-icon-button mat-fab *ngIf="isSignin" (click)="isSignin=false">
+    <button class="mat-elevation-z0" mat-icon-button mat-fab *ngIf="isSignin" (click)="toSignup()">
       <mat-icon class="rotate" svgIcon="arrow"></mat-icon>
     </button>
-    <button class="mat-elevation-z0" mat-icon-button mat-fab *ngIf="!isSignin" (click)="isSignin=true">
+    <button class="mat-elevation-z0" mat-icon-button mat-fab *ngIf="!isSignin" (click)="toSignin()">
       <mat-icon svgIcon="arrow"></mat-icon>
     </button>
   </mat-sidenav-content>
-  <mat-sidenav #signupSidenav [opened]="!isSignin" mode="side" position="end">
-    <auth-signup-form (opened)="isSignin=true" (submited)="signup($event)"></auth-signup-form>
+  <mat-sidenav #signinSidenav [opened]="isSignin" mode="side">
+    <auth-login-form (opened)="toSignup()" (submited)="signin($event)"></auth-login-form>
+  </mat-sidenav>
+  <mat-sidenav #signupSidenav opened=false mode="side" position="end">
+    <auth-signup-form (opened)="toSignin()" (submited)="signup($event)"></auth-signup-form>
   </mat-sidenav>
 </mat-sidenav-container>
