+++ libs/auth/src/lib/pages/login-view/login-view.component.scss
@@ -23,7 +23,7 @@ button {
 .mat-fab {
   height: 120px;
   width: 120px;
-  background-color: var(--iconBackground);
+  background-color: rgba(0,0,0,0.3);
 }
 
 mat-icon {
@@ -31,7 +31,7 @@ mat-icon {
 }
 
 mat-sidenav {
-  width: 700px;
+  width: 800px;
 }
 
 // large
@@ -61,3 +61,25 @@ mat-sidenav {
     width: 100%;
   }
 }
+
+
+// section {
+//   height: 100%;
+//   width: 50%;
+//   padding-top: 100px;
+// }
+
+// hr {
+//   border-right: none;
+//   border-left: solid 0.5px rgba(50, 50, 50, 0.5);
+//   height: 60%;
+// }
+
+// h1 {
+//   width: 50%;
+//   text-transform: uppercase;
+// }
+
+// form {
+//   width: 50%;
+// }
