--- libs/auth/src/lib/root/root.component.html
@@ -1,16 +1,0 @@
-<mat-sidenav-container>
-  <mat-sidenav-content [fxLayoutAlign]="align">
-    <button *ngIf="!isSignin" mat-stroked-button (click)="toSignin()">
-      open login
-    </button>
-    <button *ngIf="isSignin" mat-stroked-button (click)="toSignup()">
-      open signin
-    </button>
-  </mat-sidenav-content>
-  <mat-sidenav #signinSidenav [opened]="isSignin" mode="side">
-    <auth-login-form (opened)="toSignup()" (submited)="signin($event)"></auth-login-form>
-  </mat-sidenav>
-  <mat-sidenav #signupSidenav opened=false mode="side" position="end">
-    <auth-signup-form (opened)="toSignin()" (submited)="signup($event)"></auth-signup-form>
-  </mat-sidenav>
-</mat-sidenav-container>
